src/code2/requirements.txt
numpy>=1.22.4
sklearn>=0.0
transformers>=4.0.0
pathlib>=1.0.1
pyyaml>=6.0
scikit-learn>=1.1.1
torchvision>=0.12.0
tqdm>=4.64.0
pandas>=1.4.2
matplotlib>=3.5.2